FROMLIST: media: rockchip: Move H264 CABAC table to header file
The table will be shared with the rkvdec2 driver in following commits.

As HW formatted CABAC tables can be shared between drivers, it makes sense
to group them in an include folder.

Each driver can include the tables they need so that all tables are not
built in the kernel, but rather in each driver, if the driver is enabled.

Signed-off-by: Detlev Casanova <detlev.casanova@collabora.com>