arm64: dts: rockchip: Adjust the HDMI DDC IO driver strength for rk3588
For the RK3588 HDMI controller, the falling edge of DDC SDA and SCL almost
coincide and cannot be adjusted by HDMI registrer, resulting in poor compatibility
of DDC communication.

An improvement of the compatibility of DDC can be done by increasing the driver
strength of SCL and decreasing the driver strength of SDA to increase the slope
of the falling edge.

Signed-off-by: Andy Yan <andy.yan@rock-chips.com>